Related work items: #94307, #106566, #106862, #106877, #107179

+++ Logitude/Unifreight.Data/AmitalModel/EntityPOCOs/AmitalDataModel.GGGQ.cs
@@ -282,7 +282,7 @@ namespace Unifreight.Data.AmitalModel.EntityPOCOs
         /// There are no comments for EXPTASKTIME in the schema.
         /// </summary>
         [System.Runtime.Serialization.DataMember]
-        public virtual global::System.Nullable<long> EXPTASKTIME
+        public virtual global::System.Nullable<decimal> EXPTASKTIME
         {
             get
             {
@@ -297,7 +297,7 @@ namespace Unifreight.Data.AmitalModel.EntityPOCOs
                 }
             }
         }
-        private global::System.Nullable<long> _EXPTASKTIME;
+        private global::System.Nullable<decimal> _EXPTASKTIME;
 
     
         /// <summary>
